Task#9 - refactoring dao structure
