+++ src/main/java/com/itechart/javalab/library/dao/CrudBookDao.java
@@ -1,0 +1,20 @@
+package com.itechart.javalab.library.dao;
+
+import com.itechart.javalab.library.model.Book;
+import com.itechart.javalab.library.model.BookFilter;
+import com.itechart.javalab.library.model.Paginator;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Optional;
+
+public interface CrudBookDao {
+
+    Optional<Boolean> updateBookInfo(Book book);
+
+    boolean deleteBooks(int[] bookId);
+
+    int createBook(Book book);
+
+    void updateBookCover(Book book);
+}
