+++ src/main/java/com/itechart/javalab/library/service/impl/DefaultBookService.java
@@ -1,7 +1,9 @@
 package com.itechart.javalab.library.service.impl;
 
-import com.itechart.javalab.library.dao.BookDao;
-import com.itechart.javalab.library.dao.impl.SqlBookDao;
+import com.itechart.javalab.library.dao.CrudBookDao;
+import com.itechart.javalab.library.dao.DefaultBookDao;
+import com.itechart.javalab.library.dao.impl.SqlCrudBookDao;
+import com.itechart.javalab.library.dao.impl.SqlDefaultBookDao;
 import com.itechart.javalab.library.dto.BookDto;
 import com.itechart.javalab.library.model.Book;
 import com.itechart.javalab.library.model.BookFilter;
@@ -16,20 +18,25 @@ import java.util.Optional;
 
 public class DefaultBookService implements BookService {
 
-    private final BookDao bookDao;
+    private final CrudBookDao crudBookDao;
+    private final DefaultBookDao defaultBookDao;
+
     private final UploadFileService uploadFileService;
     private static volatile BookService instance;
 
-    private DefaultBookService(BookDao bookDao, UploadFileService uploadFileService) {
-        this.bookDao = bookDao;
-        this.uploadFileService = uploadFileService;
+    private DefaultBookService(CrudBookDao crudBookDao, DefaultBookDao defaultBookDao, UploadFileService fileService) {
+        this.crudBookDao = crudBookDao;
+        this.defaultBookDao = defaultBookDao;
+        this.uploadFileService = fileService;
     }
 
     public static BookService getInstance() {
         if (instance == null) {
             synchronized (BookService.class) {
                 if (instance == null) {
-                    instance = new DefaultBookService(SqlBookDao.getInstance(), DefaultUploadFileService.getInstance());
+                    instance = new DefaultBookService(SqlCrudBookDao.getInstance(),
+                            SqlDefaultBookDao.getInstance(),
+                            DefaultUploadFileService.getInstance());
                 }
             }
         }
@@ -38,27 +45,27 @@ public class DefaultBookService implements BookService {
 
     @Override
     public Optional<List<Book>> getBooks(Paginator paginator, boolean isAvailableOnly) {
-        return bookDao.getBooks(paginator, isAvailableOnly);
+        return defaultBookDao.getBooks(paginator, isAvailableOnly);
     }
 
     @Override
     public Optional<Integer> getNumberBooksRecords(boolean isAvailableOnly) {
-        return bookDao.getNumberBooksRecords(isAvailableOnly);
+        return defaultBookDao.getNumberBooksRecords(isAvailableOnly);
     }
 
     @Override
     public Optional<List<Book>> findBooksByParameters(Paginator paginator, BookFilter bookFilter) {
-        return bookDao.findBooksByParameters(paginator, bookFilter);
+        return defaultBookDao.findBooksByParameters(paginator, bookFilter);
     }
 
     @Override
     public Optional<Integer> getNumberFoundBooksRecords(BookFilter bookFilter) {
-        return bookDao.getNumberFoundBooksRecords(bookFilter);
+        return defaultBookDao.getNumberFoundBooksRecords(bookFilter);
     }
 
     @Override
     public Optional<Book> getBookById(int bookId) {
-        Optional<Book> bookById = bookDao.getBookById(bookId);
+        Optional<Book> bookById = defaultBookDao.getBookById(bookId);
         if (bookById.isPresent()) {
             Book book = bookById.get();
             calculateBookStatus(book);
@@ -72,7 +79,7 @@ public class DefaultBookService implements BookService {
     public Optional<Boolean> updateBookInfo(BookDto bookDto, Part part, String savePath) {
         Book book = Book.buildFrom(bookDto);
         String nameForDb = defineFileName(part, book);
-        Optional<Boolean> updateResult = bookDao.updateBookInfo(book);
+        Optional<Boolean> updateResult = crudBookDao.updateBookInfo(book);
         if (part != null && updateResult.isPresent()) {
             if (part.getSize() != 0) {
                 uploadFileService.uploadFile(savePath, part, nameForDb);
@@ -83,27 +90,27 @@ public class DefaultBookService implements BookService {
 
     @Override
     public boolean deleteBooks(int[] booksId) {
-        return bookDao.deleteBooks(booksId);
+        return crudBookDao.deleteBooks(booksId);
     }
 
     private void calculateBookStatus(Book book) {
         if (book.getInStock() > 0) {
             book.setAvailableStatus();
         } else {
-            bookDao.getEarliestDueDate(book.getId()).ifPresent(book::setUnavailableStatus);
+            defaultBookDao.getEarliestDueDate(book.getId()).ifPresent(book::setUnavailableStatus);
         }
     }
 
     @Override
     public int createBook(BookDto bookDto, Part part, String savePath) {
         Book book = Book.buildFrom(bookDto);
-        int id = bookDao.createBook(book);
+        int id = crudBookDao.createBook(book);
         book.setId(id);
         String cover = defineFileName(part, book);
         book.setCoverPath(cover);
         if (part != null && part.getSize() != 0) {
-                bookDao.updateBookCover(book);
-                uploadFileService.uploadFile(savePath, part,cover);
+            crudBookDao.updateBookCover(book);
+            uploadFileService.uploadFile(savePath, part, cover);
         }
         return id;
     }
